Module Michelson_mcmc_samplers.Make_code_sampler

Make_code_sampler produces a sampler for well-typed Michelson code. The parameters of the functor are:

a module Michelson_base implementing samplers for basic values

a module Crypto_samplers implementing samplers for pk/pkh/sk triplets

a module X containing some parameters to the Markov chain sampler:

rng_state is the mutable state that will be used during sampling

target_size specifies the size, in terms of Micheline nodes, of the terms that the sampler should try to produce

verbosity specifies how much information should be written on stdout during the sampling process.

The outcome is a michelson_code generator. The burn_in parameter specifies how much samples should be thrown away before starting to produce sample (this is used to let the underlying Markov chain reach its stationary distribution - the value should be commensurate with the target_size.

Parameters

module Michelson_base : Michelson_samplers_base.S

module Crypto_samplers : Tezos_benchmark.Crypto_samplers.Finite_key_pool_S

module X : sig ... end

Signature

val generator : burn_in:int -> Stdlib.Random.State.t -> michelson_code Stats.Gen.t

generator ~burn_in rng_state performs a burn-in phase consisting of sampling burn_in times, throwing the results away and returns a michelson term sampler. The goal of burn-in is to drive the underlying Markov chain to its stationary distribution, ie to sample terms around the specified X.target_size.