Module Michelson_mcmc_samplers.Make_data_sampler

See documentation for Make_code_sampler

Parameters

module Michelson_base : Michelson_samplers_base.S

module Crypto_samplers : Tezos_benchmark.Crypto_samplers.Finite_key_pool_S

module X : sig ... end

Signature

val generator : burn_in:int -> Stdlib.Random.State.t -> michelson_data Stats.Gen.t

generator ~burn_in rng_state performs a burn-in phase consisting of sampling burn_in times, throwing the results away and returns a michelson term sampler. The goal of burn-in is to drive the underlying Markov chain to its stationary distribution, ie to sample terms around the specified X.target_size.