Module Tezos_benchmark_020_PsParisC.Michelson_samplers

Sampling various Michelson values.

exception SamplingError of string

This module exposes a functor implementing various samplers for Michelson. These allow to sample:

types and comparable types (given a target size),

values and comparable values of a given Michelson type (given some more parameters fixed at functor instantiation time)

stacks

Note that some kind of values might not be supported. At the time of writing, the value sampler doesn't handle the following types:

Sapling transaction and states

Timelock chests and chest keys

Operations

Lambdas (ie code)

For the latter, consider using the samplers in Michelson_mcmc_samplers.

type parameters = {

base_parameters : Michelson_samplers_base.parameters;

list_size : Tezos_benchmark.Base_samplers.range;
(*
The range of the size, measured in number of elements, in which lists must be sampled.
*)

set_size : Tezos_benchmark.Base_samplers.range;
(*
The range of the size, measured in number of elements, in which sets must be sampled.
*)

map_size : Tezos_benchmark.Base_samplers.range;
(*
The range of the size, measured in number of bindings, in which maps must be sampled.
*)

}

Parameters for the Michelson samplers.

val parameters_encoding : parameters Data_encoding.t

Encoding for sampler prameters.

type type_name = [

| `TUnit

| `TInt

| `TNat

| `TSignature

| `TString

| `TBytes

| `TMutez

| `TKey_hash

| `TKey

| `TTimestamp

| `TAddress

| `TBool

| `TPair

| `TOr

| `TLambda

| `TOption

| `TList

| `TSet

| `TMap

| `TBig_map

| `TContract

| `TSapling_transaction

| `TSapling_transaction_deprecated

| `TSapling_state

| `TOperation

| `TChain_id

| `TBls12_381_g1

| `TBls12_381_g2

| `TBls12_381_fr

| `TTicket

]

module type S = sig ... end

The module type produced by the Make functor.

module Make (P : sig ... end) (Crypto_samplers : Tezos_benchmark.Crypto_samplers.Finite_key_pool_S) : S

Instantiate a module of type S.

module Internal_for_tests : sig ... end