Module Tezos_benchmark_020_PsParisC.Type_helpers

Type conversion helpers

exception Type_helpers_error of string

Exception raised in case an error occurs in this module.

val michelson_type_list_to_ex_stack_ty : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Script.expr list -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.t -> Tezos_protocol_020_PsParisC.Protocol.Script_ir_translator.ex_stack_ty

michelson_type_list_to_ex_stack_ty converts a list of types in Micheline form to a stack type in IR form.

raises Type_helpers_error
if parsing the Michelson type fails.

val michelson_type_to_ex_ty : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Script.expr -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.t -> Tezos_protocol_020_PsParisC.Protocol.Script_typed_ir.ex_ty

michelson_type_to_ex_ty ty ctxt parses the type ty.

raises Type_helpers_error
if an error arises during parsing.

val stack_type_to_michelson_type_list : Tezos_benchmark_type_inference_020_PsParisC.Type.Stack.t -> Tezos_protocol_020_PsParisC.Protocol.Script_repr.expr list

stack_type_to_michelson_type_list converts a Mikhailsky stack type to a stack represented as a list of Micheline expressions, each element denoting a type on the stack type.

raises Type_helpers_error
if the stack type contains variables.

val base_type_to_ex_ty : Tezos_benchmark_type_inference_020_PsParisC.Type.Base.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.t -> Tezos_protocol_020_PsParisC.Protocol.Script_typed_ir.ex_ty

base_type_to_ex_ty converts a Mikhailsky type to a Michelson one.